refactor(playback): 💀 remove playback feature
